feat: move to `projectView` package
