+++ libs/ide/rust-extra/src/main/kotlin/ai/hash/rustextra/projectView/ModuleFileNode.kt
@@ -1,13 +1,11 @@
-package ai.hash.rustextra
+package ai.hash.rustextra.projectView
 
 import com.intellij.ide.projectView.PresentationData
 import com.intellij.ide.projectView.ViewSettings
 import com.intellij.ide.projectView.impl.nodes.PsiFileNode
-import com.intellij.ide.util.treeView.AbstractTreeNode
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import com.intellij.ui.JBColor
-import java.awt.Color
 
 class ModuleFileNode(project: Project, value: PsiFile, viewSettings: ViewSettings) : PsiFileNode(project, value, viewSettings) {
