+++ libs/ide/rust-extra/src/main/resources/META-INF/plugin.xml
@@ -25,6 +25,6 @@
     <!-- Extension points defined by the plugin.
          Read more: https://plugins.jetbrains.com/docs/intellij/plugin-extension-points.html -->
     <extensions defaultExtensionNs="com.intellij">
-        <treeStructureProvider implementation="ai.hash.rustextra.ModuleGroup"/>
+        <treeStructureProvider implementation="ai.hash.rustextra.projectView.ModuleGroup"/>
     </extensions>
 </idea-plugin>
